azerty

+++ src/main/java/com/test/springWebApp/repositories/AuthorRepositary.java
@@ -4,5 +4,5 @@ import com.test.springWebApp.model.Author;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface AuthorRepositary extends CrudRepository<Author, Integer> {
+public interface AuthorRepositary extends CrudRepository<Author, Long> {
 }
