+++ src/main/java/com/test/springWebApp/repositories/BookRepository.java
@@ -5,5 +5,5 @@ import com.test.springWebApp.model.Book;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 }
