+++ src/main/java/com/test/springWebApp/repositories/PublisherRepository.java
@@ -5,5 +5,5 @@ import com.test.springWebApp.model.Publisher;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface PublisherRepository extends CrudRepository<Publisher, Integer> {
+public interface PublisherRepository extends CrudRepository<Publisher, Long> {
 }
